move SidebarTemplate to common
