+++ src/common/sidebar/index.tsx
@@ -1,1 +1,2 @@
 export { SidebarContainer as ContentSidebar } from './containers/SidebarContainer';
+export { SidebarTemplate } from './containers/SidebarTemplate';
